fixed bug in segment interpretation of 2opt, validated with distance_table v ersion

+++ distance_table/types.hh
@@ -1,7 +1,7 @@
 #ifndef __TYPES_H__
 #define __TYPES_H__
 
-typedef float dtype;
+typedef double dtype;
 typedef long int cost_t;
 typedef dtype (*distanceFunction)(dtype*,dtype*,int,int);
